requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.6.4
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.0
alembic==1.16.1
pytest
pytest-asyncio
httpx
typing_extensions>=4.0.0